+++ packages/api-reference/src/components/Content/Content.vue
@@ -3,12 +3,12 @@ import { useActiveEntities, useWorkspace } from '@scalar/api-client/store'
 import { RequestAuth } from '@scalar/api-client/views/Request/RequestSection/RequestAuth'
 import { ScalarErrorBoundary } from '@scalar/components'
 import { getSlugUid } from '@scalar/oas-utils/transforms'
+import type { OpenAPIV3_1 } from '@scalar/openapi-types'
 import type { Spec } from '@scalar/types/legacy'
 import { computed } from 'vue'
 
 import { BaseUrl } from '@/features/BaseUrl'
 import { useConfig } from '@/hooks/useConfig'
-import { useSidebar } from '@/hooks/useSidebar'
 import { getModels, hasModels } from '@/libs/openapi'
 
 import { ClientLibraries } from './ClientLibraries'
@@ -16,10 +16,10 @@ import { Introduction } from './Introduction'
 import { Loading } from './Lazy'
 import { Models, ModelsAccordion } from './Models'
 import { TagList } from './Tag'
-import { Webhooks } from './Webhooks'
 
 const props = withDefaults(
   defineProps<{
+    document: OpenAPIV3_1.Document
     parsedSpec: Spec
     layout?: 'modern' | 'classic'
   }>(),
@@ -29,7 +29,6 @@ const props = withDefaults(
 )
 
 const config = useConfig()
-const { hideModels } = useSidebar()
 const { collections, securitySchemes, servers } = useWorkspace()
 const {
   activeCollection: _activeCollection,
@@ -91,9 +90,8 @@ const introCardsSlot = computed(() =>
       :server="activeServer" />
 
     <Introduction
-      v-if="parsedSpec?.info?.title || parsedSpec?.info?.description"
-      :info="parsedSpec.info"
-      :parsedSpec="parsedSpec">
+      v-if="document?.info?.title || document?.info?.description"
+      :document="document">
       <template #[introCardsSlot]>
         <ScalarErrorBoundary>
           <div
@@ -160,11 +158,26 @@ const introCardsSlot = computed(() =>
         :tags="parsedSpec.tags" />
     </template>
 
-    <template v-if="parsedSpec.webhooks">
-      <Webhooks :webhooks="parsedSpec.webhooks" />
+    <!-- Webhooks -->
+    <template v-if="parsedSpec.webhooks?.length && activeCollection">
+      <TagList
+        id="webhooks"
+        :collection="activeCollection"
+        :layout="layout"
+        :schemas="getModels(parsedSpec)"
+        :server="activeServer"
+        :spec="parsedSpec"
+        :tags="[
+          {
+            name: 'Webhooks',
+            description: '',
+            operations: parsedSpec.webhooks,
+          },
+        ]">
+      </TagList>
     </template>
 
-    <template v-if="hasModels(parsedSpec) && !hideModels">
+    <template v-if="hasModels(parsedSpec) && !config.hideModels">
       <ModelsAccordion
         v-if="layout === 'classic'"
         :schemas="getModels(parsedSpec)" />
@@ -191,7 +204,6 @@ const introCardsSlot = computed(() =>
 .introduction-card {
   display: flex;
   flex-direction: column;
-  background: var(--scalar-background-1);
 }
 .introduction-card-item {
   display: flex;