+++ packages/api-reference/src/components/Content/Tag/TagList.vue
@@ -7,14 +7,15 @@ import { computed } from 'vue'
 
 import { Lazy } from '@/components/Content/Lazy'
 import { Operation } from '@/features/Operation'
-import { operationIdParams } from '@/features/traverse-schema'
+import { useSidebar } from '@/features/sidebar'
 import { useNavState } from '@/hooks/useNavState'
-import { useSidebar } from '@/hooks/useSidebar'
 
 import TagAccordion from './TagAccordion.vue'
 import TagSection from './TagSection.vue'
 
 const { collection, tags, spec, layout, server } = defineProps<{
+  /** Just to set the id for webhooks, for now */
+  id?: string
   collection: Collection
   tags: TagType[]
   spec: Spec
@@ -23,7 +24,7 @@ const { collection, tags, spec, layout, server } = defineProps<{
   schemas?: Record<string, OpenAPIV3_1.SchemaObject> | unknown
 }>()
 
-const { getOperationId, getTagId, hash } = useNavState()
+const { getTagId, hash } = useNavState()
 const { collapsedSidebarItems } = useSidebar()
 
 const tagLayout = computed(() =>
@@ -47,26 +48,25 @@ const isLazy = (index: number) =>
 <template>
   <Lazy
     v-for="(tag, index) in tags"
-    :id="getTagId(tag)"
-    :key="getTagId(tag)"
+    :id="id || getTagId(tag)"
+    :key="id || getTagId(tag)"
     :isLazy="isLazy(index)">
     <Component
       :is="tagLayout"
-      :id="getTagId(tag)"
+      :id="id || getTagId(tag)"
       :collection="collection"
       :spec="spec"
       :tag="tag">
       <Lazy
         v-for="(operation, operationIndex) in tag.operations"
-        :id="getOperationId(operationIdParams(operation), tag)"
-        :key="`${operation.httpVerb}-${operation.operationId}`"
+        :id="operation.id"
+        :key="operation.id"
         :isLazy="
           isLazy(index) ||
           (collapsedSidebarItems[getTagId(tag)] && operationIndex > 0)
         ">
         <ScalarErrorBoundary>
           <Operation
-            :id="getOperationId(operationIdParams(operation), tag)"
             :collection="collection"
             :layout="layout"
             :schemas="schemas"