+++ packages/api-reference/src/components/MobileHeader.vue
@@ -1,25 +1,21 @@
 <script setup lang="ts">
 import { ScalarIconButton } from '@scalar/components'
 
-import { useSidebar } from '@/hooks/useSidebar'
+import { useSidebar } from '@/features/sidebar'
+
+const { isSidebarOpen } = useSidebar()
 
 defineProps<{
-  open?: boolean
+  breadcrumb: string
 }>()
-
-defineEmits<{
-  (event: 'update:open', open: boolean): void
-}>()
-
-const { breadcrumb } = useSidebar()
 </script>
 <template>
   <div class="references-mobile-header t-doc__header">
     <ScalarIconButton
-      :icon="open ? 'Close' : 'Menu'"
-      :label="open ? 'Close Menu' : 'Open Menu'"
+      :icon="isSidebarOpen ? 'Close' : 'Menu'"
+      :label="isSidebarOpen ? 'Close Menu' : 'Open Menu'"
       size="md"
-      @click="$emit('update:open', !open)" />
+      @click="isSidebarOpen = !isSidebarOpen" />
     <span class="references-mobile-breadcrumbs">{{ breadcrumb }}</span>
     <div class="references-mobile-header-actions">
       <slot name="actions" />