+++ packages/api-reference/src/features/Search/useSearchIndex.test.ts
@@ -5,6 +5,7 @@ import { parse } from '@/helpers/parse'
 import { createEmptySpecification } from '@/libs/openapi'
 import { dereference } from '@scalar/openapi-parser'
 import type { OpenAPIV3_1 } from '@scalar/openapi-types'
+import type { TraversedEntry } from '@/features/traverse-schema'
 import { useSearchIndex } from './useSearchIndex'
 
 // Mock the useConfig hook
@@ -34,7 +35,31 @@ describe('useSearchIndex', () => {
       }),
     )
 
-    const specification = await parse(dereferencedDocument as OpenAPIV3_1.Document)
+    const items: TraversedEntry[] = [
+      {
+        id: 'tag/default',
+        isGroup: true,
+        title: 'Default',
+        tag: {
+          name: 'default',
+          description: 'Default tag',
+        },
+        children: [
+          {
+            id: 'tag/default/foobar',
+            method: 'get',
+            path: '/foobar',
+            title: 'Get request',
+            operation: {
+              summary: 'Get request',
+              description: 'Get request description',
+            },
+          },
+        ],
+      },
+    ]
+
+    const specification = await parse(dereferencedDocument as OpenAPIV3_1.Document, items)
 
     const { searchText, fuseSearch, searchResultsWithPlaceholderResults } = useSearchIndex({
       specification: toRef(specification),