+++ packages/api-reference/src/hooks/useNavState.test.ts
@@ -4,7 +4,6 @@ import { apiReferenceConfigurationSchema } from '@scalar/types/api-reference'
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 import { computed, inject, ref } from 'vue'
 
-import { operationIdParams } from '@/features/traverse-schema'
 import { useNavState } from './useNavState'
 
 declare global {
@@ -115,7 +114,7 @@ describe('useNavState', () => {
 
     it('should generate operation ID', () => {
       const operation = {
-        httpVerb: 'GET',
+        method: 'get',
         path: '/test',
       } as const
       const parentTag = {
@@ -123,7 +122,7 @@ describe('useNavState', () => {
         description: 'Test Description',
         operations: [],
       } satisfies Tag
-      expect(navState.getOperationId(operationIdParams(operation), parentTag)).toBe('tag/test-tag/get/test')
+      expect(navState.getOperationId(operation, parentTag)).toBe('tag/test-tag/get/test')
     })
 
     it('should generate webhook ID', () => {