+++ packages/openapi-types/src/helpers/index.ts
@@ -1,0 +1,1 @@
+export { isDereferenced } from './is-dereferenced'