+++ packages/scripts/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @scalar/scripts
 
+## 0.0.10
+
+### Patch Changes
+
+- @scalar/components@0.14.9
+
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies [2974dd6]
+- Updated dependencies [221e35f]
+- Updated dependencies [4156f1d]
+- Updated dependencies [177438d]
+  - @scalar/oas-utils@0.4.5
+  - @scalar/components@0.14.8
+
 ## 0.0.8
 
 ### Patch Changes