+++ packages/themes/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @scalar/themes
 
+## 0.13.3
+
+### Patch Changes
+
+- Updated dependencies [221e35f]
+  - @scalar/types@0.2.3
+
 ## 0.13.2
 
 ### Patch Changes