+++ projects/scalar-app/CHANGELOG.md
@@ -1,5 +1,29 @@
 # scalar-app
 
+## 0.1.205
+
+### Patch Changes
+
+- Updated dependencies [a5d010c]
+- Updated dependencies [b12cd8e]
+  - @scalar/api-client@2.5.7
+  - @scalar/components@0.14.9
+
+## 0.1.204
+
+### Patch Changes
+
+- 4156f1d: Expand workspace store integration
+- Updated dependencies [282129b]
+- Updated dependencies [a5b399d]
+- Updated dependencies [4156f1d]
+- Updated dependencies [711b435]
+- Updated dependencies [177438d]
+  - @scalar/api-client@2.5.6
+  - @scalar/components@0.14.8
+  - @scalar/import@0.4.10
+  - @scalar/themes@0.13.3
+
 ## 0.1.203
 
 ### Patch Changes